making final edits, move main method
