+++ ClientSideKeyVaultKeyToMicrosoftManagedKey/migration.py
@@ -51,9 +51,9 @@ def get_keyvault_key(credential, k_client):
     return keyvault_key
 
 
-def download_blob(blob_name, container_client):
+def download_blob(blob_name, container_client, kvk, credential):
     print("\nDownloading and decrypting blob...")
-    kek = KeyWrapper(kvk, credentials)
+    kek = KeyWrapper(kvk, credential)
     container_client.key_encryption_key = kek
     decrypted_message = open("decryptedcontentfile.txt", "wb+")
     decrypted_message.write(container_client.get_blob_client(blob_name).download_blob().content_as_bytes())
@@ -75,13 +75,13 @@ def upload_blob(blob_service_client, cont_name, blob_name):
     # upload using microsoft managed encryption-scope
     print("\nPerforming server side encryption with microsoft managed key...")
     # access container and specified blob name
-    blob_client = bs_client.get_blob_client(container=cont_name, blob=cfg.migrated_blob_name)
+    blob_client = blob_service_client.get_blob_client(container=cont_name, blob=cfg.migrated_blob_name)
     # upload contents to that blob with microsoft managed key for server side encryption
     blob_client.upload_blob(file_content, encryption_scope=cfg.server_managed_encryption_scope,
                             blob_type=b_type, overwrite=True)
 
 
-if __name__ == "__main__":
+def main():
     # credential required to access client account
     credentials = ClientSecretCredential(cfg.TENANT_ID, cfg.CLIENT_ID,
                                          cfg.CLIENT_SECRET)
@@ -93,6 +93,10 @@ if __name__ == "__main__":
     cont_client = bs_client.get_container_client(cfg.cont_name)
 
     # call to methods
-    kvk = get_keyvault_key(credentials, key_client)
-    download_blob(cfg.blob_name, cont_client)
+    key_vault_key = get_keyvault_key(credentials, key_client)
+    download_blob(cfg.blob_name, cont_client, key_vault_key, credentials)
     upload_blob(bs_client, cfg.cont_name, cfg.blob_name)
+
+
+if __name__ == "__main__":
+    main()
