+++ ClientSideLocalKeyToCustomerManagedKey/setup/blobExample.txt
@@ -1,1 +1,1 @@
-test file
+Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec tincidunt sem vitae dolor efficitur fermentum. Nulla nec nunc porta, tempus quam vel, condimentum ipsum. Cras vehicula ut libero ac convallis. Nullam fringilla at magna sed lacinia. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia curae; Nulla eu aliquam leo. Aliquam et enim neque. Quisque dolor libero, pharetra eu eros quis, sodales iaculis justo. Aenean pharetra aliquet ex, sed finibus quam suscipit vitae. Lorem ipsum dolor sit amet, consectetur adipiscing elit.
